Don't consider the multiplication factor when choosing fnum/block
Initially, the intention was to adjust the fnum/block based on the
multiplication factor, so that all instruments could sound in the same
octave when played.

However, which octave the instrument sounds in can't be determined by
only looking at a single multiplication factor. Depending on the depth
of the modulation, one or the other may have the greater influence.

Instead, keep the fnum/block consistent for the keyboard.